refactor: Store, Location을 mission 도메인으로 통합
