+++ src/main/java/com/example/umc10th/domain/mission/entity/Mission.java
@@ -1,7 +1,6 @@
 package com.example.umc10th.domain.mission.entity;
 
 import com.example.umc10th.global.entity.BaseEntity;
-import com.example.umc10th.domain.store.entity.Store;
 import jakarta.persistence.*;
 import lombok.*;
