+++ src/main/java/com/example/umc10th/domain/mission/repository/StoreRepository.java
@@ -1,6 +1,6 @@
-package com.example.umc10th.domain.store.repository;
+package com.example.umc10th.domain.mission.repository;
 
-import com.example.umc10th.domain.store.entity.Store;
+import com.example.umc10th.domain.mission.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StoreRepository extends JpaRepository<Store, Long> {
